+++ kaithem/src/tagpoints.py
@@ -791,6 +791,8 @@ class _TagPoint(virtualresource.VirtualResource):
                     return self.meterWidget.render()
                 elif hasattr(self, "spanWidget"):
                     return self.spanWidget.render()
+                else:
+                    return "Binary Tagpoint"
             except Exception as e:
                 return str(e)
 
@@ -978,7 +980,15 @@ class _TagPoint(virtualresource.VirtualResource):
 
             # Val override last, in case it triggers an alarm
             # Convert to string for consistent handling, the config engine things anything that looks like a number, is.
-            overrideValue = str(data.get('overrideValue', '')).strip()
+            overrideValue = str(data.get('overrideValue', '')).strip().replace(" ","")
+
+            if self.type == "binary":
+                try:
+                    overrideValue = bytes.fromhex(overrideValue)
+                except:
+                    logging.exception("Bad hex in tag override")
+                    overrideValue = b''
+
             if overrideValue:
                 if overrideValue.startswith("="):
                     self.kweb_manualOverrideClaim = createGetterFromExpression(
@@ -1907,35 +1917,35 @@ class _ObjectTagPoint(_TagPoint):
             self.lock.release()
 
 
-# class _BinaryTagPoint(_TagPoint):
-#     defaultData: bytes = b''
-#     type = 'binary'
+class _BinaryTagPoint(_TagPoint):
+    defaultData: bytes = b''
+    type = 'binary'
 
-#     @typechecked
-#     def __init__(self, name: str):
-#         self.guiLock = threading.Lock()
+    @typechecked
+    def __init__(self, name: str):
+        self.guiLock = threading.Lock()
 
-#         self.validate = None
-#         _TagPoint.__init__(self, name)
+        self.validate = None
+        _TagPoint.__init__(self, name)
 
-#     def processValue(self, value):
-#         # Functions are special valid types of value.
-#         # They are automatically resolved.
-#         if callable(value):
-#             value = value()
+    def processValue(self, value):
+        # Functions are special valid types of value.
+        # They are automatically resolved.
+        if callable(value):
+            value = value()
 
-#         if isinstance(value, bytes):
-#             value = value
-#         else:
-#             value=bytes(value)
+        if isinstance(value, bytes):
+            value = value
+        else:
+            value=bytes(value)
 
-#         if self.validate:
-#             value = self.validate(value)
+        if self.validate:
+            value = self.validate(value)
 
-#         return value
+        return value
 
-#     def filterValue(self, v):
-#         return v
+    def filterValue(self, v):
+        return v
 
 
 
@@ -2168,3 +2178,4 @@ def createGetterFromExpression(e, t, priority=98):
 Tag = _NumericTagPoint.Tag
 ObjectTag = _ObjectTagPoint.Tag
 StringTag = _StringTagPoint.Tag
+BinaryTag = _BinaryTagPoint.Tag